chore: code formatting with forge fmt

+++ contracts/mocks/ReentrancyAttacker.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity =0.8.25;
 
-import { IDiamondDao } from "../interfaces/IDiamondDao.sol";
-import { IDiamondDaoLowMajority } from "../interfaces/IDiamondDaoLowMajority.sol";
+import {IDiamondDao} from "../interfaces/IDiamondDao.sol";
+import {IDiamondDaoLowMajority} from "../interfaces/IDiamondDaoLowMajority.sol";
 
 contract ReentrancyAttacker {
     IDiamondDao public dao;
@@ -29,7 +29,7 @@ contract ReentrancyAttacker {
         dao.execute(proposalId);
     }
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return address(this).balance;
     }
 }
@@ -55,12 +55,12 @@ contract ReentrancyAttackerLowMajority {
 
         if (calls != 5) {
             dao.execute(1, _targets, _values, _calldata);
-        
+
             ++calls;
         }
     }
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return address(this).balance;
     }
 }
